requirements.txt
pytest>=9.0
argon2-cffi>=23.1.0
cryptography>=42.0.0
